moving pre-registration
