+++ services/web/server/src/simcore_service_webserver/login/_controller/rest/twofa.py
@@ -13,16 +13,16 @@ from ....products import products_web
 from ....products.models import Product
 from ....session.access_policies import session_access_required
 from ... import _twofa_service
-from ..._constants import (
+from ..._login_repository_legacy import AsyncpgStorage, get_plugin_storage
+from ..._login_service import envelope_response
+from ..._models import InputSchema
+from ...constants import (
     CODE_2FA_EMAIL_CODE_REQUIRED,
     CODE_2FA_SMS_CODE_REQUIRED,
     MSG_2FA_CODE_SENT,
     MSG_EMAIL_SENT,
     MSG_UNKNOWN_EMAIL,
 )
-from ..._login_repository_legacy import AsyncpgStorage, get_plugin_storage
-from ..._login_service import envelope_response
-from ..._models import InputSchema
 from ...settings import LoginSettingsForProduct, get_plugin_settings
 from ._rest_exceptions import handle_rest_requests_exceptions
