+++ services/web/server/src/simcore_service_webserver/login_account/__init__.py
@@ -1,0 +1,3 @@
+"""
+Account management
+"""
