+++ services/web/server/tests/unit/with_dbs/03/login/test_login_auth.py
@@ -23,7 +23,7 @@ from settings_library.utils_session import DEFAULT_SESSION_COOKIE_NAME
 from simcore_postgres_database.models.users import UserRole
 from simcore_service_webserver.constants import APP_SETTINGS_KEY
 from simcore_service_webserver.db.models import UserStatus
-from simcore_service_webserver.login._constants import (
+from simcore_service_webserver.login.constants import (
     MSG_ACTIVATION_REQUIRED,
     MSG_LOGGED_IN,
     MSG_UNKNOWN_EMAIL,
