+++ services/web/server/tests/unit/with_dbs/03/login/test_login_change_email.py
@@ -8,7 +8,7 @@ from pytest_simcore.helpers.assert_checks import assert_status
 from pytest_simcore.helpers.webserver_login import LoggedUser, NewUser, parse_link
 from servicelib.aiohttp import status
 from simcore_service_webserver.constants import INDEX_RESOURCE_NAME
-from simcore_service_webserver.login._constants import (
+from simcore_service_webserver.login.constants import (
     MSG_CHANGE_EMAIL_REQUESTED,
     MSG_LOGGED_IN,
     MSG_LOGGED_OUT,
