+++ services/web/server/tests/unit/with_dbs/03/login/test_login_change_password.py
@@ -9,7 +9,7 @@ from pytest_simcore.helpers.assert_checks import assert_status
 from pytest_simcore.helpers.webserver_login import LoggedUser
 from servicelib.aiohttp import status
 from servicelib.rest_responses import unwrap_envelope
-from simcore_service_webserver.login._constants import (
+from simcore_service_webserver.login.constants import (
     MSG_LOGGED_IN,
     MSG_PASSWORD_CHANGED,
     MSG_PASSWORD_MISMATCH,
