+++ services/web/server/tests/unit/with_dbs/03/login/test_login_twofa.py
@@ -21,11 +21,6 @@ from servicelib.utils_secrets import generate_passcode
 from simcore_postgres_database.models.products import ProductLoginSettingsDict, products
 from simcore_service_webserver.application_settings import ApplicationSettings
 from simcore_service_webserver.db.models import UserStatus
-from simcore_service_webserver.login._constants import (
-    CODE_2FA_SMS_CODE_REQUIRED,
-    MSG_2FA_UNAVAILABLE,
-    MSG_LOGGED_IN,
-)
 from simcore_service_webserver.login._login_repository_legacy import AsyncpgStorage
 from simcore_service_webserver.login._twofa_service import (
     _do_create_2fa_code,
@@ -34,6 +29,11 @@ from simcore_service_webserver.login._twofa_service import (
     get_2fa_code,
     get_redis_validation_code_client,
     send_email_code,
+)
+from simcore_service_webserver.login.constants import (
+    CODE_2FA_SMS_CODE_REQUIRED,
+    MSG_2FA_UNAVAILABLE,
+    MSG_LOGGED_IN,
 )
 from simcore_service_webserver.products import products_web
 from simcore_service_webserver.products.errors import UnknownProductError
