+++ services/web/server/tests/unit/with_dbs/03/test_session_access_policies.py
@@ -14,7 +14,7 @@ from pytest_simcore.helpers.typing_env import EnvVarsDict
 from servicelib.aiohttp import status
 from simcore_service_webserver.application_settings import ApplicationSettings
 from simcore_service_webserver.constants import APP_SETTINGS_KEY
-from simcore_service_webserver.login._constants import (
+from simcore_service_webserver.login.constants import (
     MAX_2FA_CODE_RESEND,
     MAX_2FA_CODE_TRIALS,
     MSG_UNAUTHORIZED_LOGIN_2FA,
